refactor: apply naming conventions to file names
