+++ src/auth/dto/res/consent-required-error.dto.ts
