+++ src/move-out/types/inspection-target.type.ts
